# perf: Add foreign key indexes to tables