With the latest release of `better-sqlite3@11.5` the `SQLite3` version was bumped to `3.47.0` ([changes](https://sqlite.org/releaselog/3_47_0.html)).

> No attempt is made to create [automatic indexes](https://sqlite.org/optoverview.html#autoindex) on a column that is known to be non-selective because of its use in other indexes that have been analyzed.

this seems to be the most important change for CAP applications as we create many `join` conditions using `associations` and `compositions` which `SQLite3` tries to process to the best of its ability by creating temporary automatic indexes, but with this optimization the likeliness of these indexes being created is greatly reduced. Causing some of the more complex join queries to take `5000x` longer then when using `better-sqlite@11.4`. This PR tries to mitigate the effect of the reduced number of automatically generated indexes by using the model definition to provide explicit indexes for these modeled join condition sources / targets.

An example of the impact that providing explicit indexes has on complex SFlight requests:

```
analytics/Bookings?$expand=to_Travel($expand=CurrencyCode,TravelStatus,to_Agency,to_Customer,to_Booking),to_Carrier,to_Flight&$top=30
```

|   | sqlite | postgres | hana |
| --- | --- | --- | --- |
| before | ~50 req/s | ~50 req/s | ~140 req/s |
| after | ~130 req/s | ~110 req/s | ~140 req/s |

So while it doesn't have an impact on the HANA performance it provides additional optimization information for `SQLite` and `Postgres`. Closing the gap that `SQLite` and `Postgres` always have had when being compared with `HANA` for complex join heavy queries.